update code so that gamma FW day algorithm is used

+++ src/CoefGen.cpp
@@ -16,7 +16,7 @@ CoefGen::CoefGen(Rcpp::NumericMatrix& U, Rcpp::List& gamma_specs, int n_samp) :
     m_vals_rcpp {Rcpp::NumericVector(Rcpp::no_init(gamma_specs.size() * n_samp))},
     m_vals {m_vals_rcpp.begin()},
     m_n_psi {0},
-    m_n_gamma {gamma_specs.size()}
+    m_n_gamma {static_cast<int>(gamma_specs.size())}
 {
 }
